[FIX] Refactoring
