--- src/main/java/com/example/demo/entity/mapping/UserInfo.java
@@ -1,6 +1,0 @@
-/*
-package com.example.demo.entity.mapping;
-
-public class UserInfo {
-}
-*/
