--- src/main/java/com/example/demo/entity/relations/UserRelation.java
@@ -1,6 +1,0 @@
-/*
-package com.example.demo.entity.relations;
-
-public class UserRelation {
-}
-*/
