+++ src/main/java/com/example/demo/repository/PaymentRepository.java
@@ -1,6 +1,6 @@
 package com.example.demo.repository;
 
-import com.example.demo.entity.Payment;
+import com.example.demo.entity.Payment.Payment;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
