+++ src/main/java/com/example/demo/repository/UserPaymentRepository.java
@@ -1,6 +1,6 @@
 package com.example.demo.repository;
 
-import com.example.demo.entity.UserPayment;
+import com.example.demo.entity.Payment.UserPayment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
