+++ src/main/java/com/example/demo/service/general/impl/KakaoPayServiceImpl.java
@@ -8,8 +8,8 @@ import com.example.demo.domain.converter.KakaoPayConverter;
 import com.example.demo.domain.dto.Payment.ApproveResponse;
 import com.example.demo.domain.dto.Payment.PaymentRequest;
 import com.example.demo.domain.dto.Payment.ReadyResponse;
-import com.example.demo.entity.Payment;
-import com.example.demo.entity.UserPayment;
+import com.example.demo.entity.Payment.Payment;
+import com.example.demo.entity.Payment.UserPayment;
 import com.example.demo.repository.PaymentRepository;
 import com.example.demo.repository.UserPaymentRepository;
 import com.example.demo.service.general.KakaoPayService;
