+++ src/main/java/com/example/demo/service/general/impl/UserPaymentServiceImpl.java
@@ -6,8 +6,8 @@ import com.example.demo.base.status.ErrorStatus;
 import com.example.demo.base.status.SuccessStatus;
 import com.example.demo.domain.converter.UserPaymentConverter;
 import com.example.demo.domain.dto.Payment.*;
-import com.example.demo.entity.Payment;
-import com.example.demo.entity.UserPayment;
+import com.example.demo.entity.Payment.Payment;
+import com.example.demo.entity.Payment.UserPayment;
 import com.example.demo.repository.PaymentRepository;
 import com.example.demo.repository.UserPaymentRepository;
 import com.example.demo.service.general.UserPaymentService;
